Chess library handles clock setup and management
The updated chess library sends the clock specs when given time limits.
The GameState enum is for forcing the library to send a new clock after
the first move of the game. The parameter remaining_moves is
set to 10000 to force the level command to be sent to the engine.